+++ scripts/managers/build_manager.gd
@@ -3,10 +3,7 @@ class_name BuildManager
 var constants: LandUse
 var facility_functions: Facilities.FacilityFunctions = Facilities.FacilityFunctions.new()
 
-var network: RoadNetwork
-var resources: ResourceManager
-var planet_data = null
-var taxes: TaxManager
+var planet = null
 var parent_node: Spatial
 
 var city_names: Dictionary = {}
@@ -14,12 +11,9 @@ var city_names: Dictionary = {}
 # warning-ignore:shadowed_variable
 # warning-ignore:shadowed_variable
 # warning-ignore:shadowed_variable
-func _init(consts: LandUse, net: RoadNetwork, resources: ResourceManager, planet_data, taxes: TaxManager, node: Spatial):
+func _init(consts: LandUse, planet, node: Spatial):
 	self.constants = consts
-	self.network = net
-	self.resources = resources
-	self.planet_data = planet_data
-	self.taxes = taxes
+	self.planet = planet
 	self.parent_node = node
 
 
@@ -31,17 +25,17 @@ func add_road(path: Array, road_type: int):
 	
 	var warn = null
 	for i in range(path.size()-1):
-		if taxes.budget < sum_cost + Roads.ROAD_COSTS[road_type]:
+		if planet.taxes.budget < sum_cost + Roads.ROAD_COSTS[road_type]:
 			warn = "Road not completed - not enough money!"
 			break
 		
 		var p1 = path[i]
 		var p2 = path[i+1]
-		if not network.points_connected(p1, p2):
-			network.connect_points(p1, p2, Roads.ROAD_CAPACITY[road_type])
+		if not planet.roads.points_connected(p1, p2):
+			planet.roads.connect_points(p1, p2, Roads.ROAD_CAPACITY[road_type])
 			sum_cost += Roads.ROAD_COSTS[road_type]
 	
-	taxes.budget -= sum_cost
+	planet.taxes.budget -= sum_cost
 	
 	return warn
 
@@ -53,20 +47,20 @@ func remove_road(path: Array) -> bool:
 	for i in range(path.size()-1):
 		var p1 = path[i]
 		var p2 = path[i+1]
-		if network.points_connected(p1, p2):
-			network.disconnect_points(p1, p2)
+		if planet.roads.points_connected(p1, p2):
+			planet.roads.disconnect_points(p1, p2)
 	
 	return true
 
 
 func grow_city(city: City):
 	var cost = Cities.city_growth_cost(city.radius)
-	if cost > taxes.budget:
+	if cost > planet.taxes.budget:
 		return "Not enough money to grow city (requires %d)" % cost
 	
 	city.radius += 1
-	city.update_cells(planet_data)
-	taxes.budget -= cost
+	city.update_cells(planet.planet_data)
+	planet.taxes.budget -= cost
 	
 	return null
 
@@ -80,36 +74,36 @@ func add_facility(type: String, location: int, name: String):
 		print("WARNING: no scene resource found for %s" % type)
 		return [null, "WARNING: no scene resource found for %s" % type]
 	
-	if network.has_facility(location) or planet_data.get_node(location).is_occupied:
+	if planet.roads.has_facility(location) or planet.planet_data.get_node(location).is_occupied:
 		return [null, "Location already occupied"]
 	
 	var costs = Facilities.FACILITY_COSTS[type]
-	if costs > taxes.budget:
+	if costs > planet.taxes.budget:
 		return [null, "Not enough money (requires %d)" % costs]
 	
-	if not facility_functions.can_build(type, planet_data, location):
+	if not facility_functions.can_build(type, planet.planet_data, location):
 		return [null, "Can't build this facility here"]
 	
 	var facility: Facility = load(Facilities.FACILITY_SCENES[type]).instance()
 	
-	facility.init(location, planet_data, type)
+	facility.init(location, planet.planet_data, type)
 	
-	taxes.budget -= costs
+	planet.taxes.budget -= costs
 	
 	return [add_facility_scene(facility, name), null]
 
 
 func add_facility_scene(facility: Facility, name: String):
-	network.add_facility(facility.node_id, facility)
+	planet.roads.add_facility(facility.node_id, facility)
 	
 	parent_node.add_child(facility)
 	
-	var info = planet_data.get_node(facility.node_id)
+	var info = planet.planet_data.get_node(facility.node_id)
 	facility.name = name
 	facility.translation = info.position
 	facility.look_at(2 * info.position, Vector3.UP)
 	
-	facility.on_ready(planet_data)
+	facility.on_ready(planet.planet_data)
 	
 	if facility is City:
 		city_names[facility.name] = true
@@ -117,11 +111,30 @@ func add_facility_scene(facility: Facility, name: String):
 	return facility
 
 
+func remove_facility(facility: Facility):
+	if facility is City:
+		var city: City = facility as City
+		if city.population() > 1 or not city.land_use.empty():
+			return "Can't remove city with population > 1\nor active land use!"
+	
+	if facility.city_node_id >= 0:
+		var owner: City = planet.roads.facilities[facility.city_node_id] as City
+		owner.remove_facility(facility.node_id)
+	
+	planet.roads.remove_facility(facility.node_id)
+	facility.removed(planet)
+	
+	parent_node.remove_child(facility)
+	facility.queue_free()
+	
+	return null
+
+
 # Use land_use = LandUse.LU_NONE to ignore specific requirements
 func can_set_land_use(city: City, node: int, land_use: int, consider_veg_res: bool = false):
-	if network.is_road(node) \
-			or network.has_facility(node) \
-			or planet_data.get_node(node).is_occupied:
+	if planet.roads.is_road(node) \
+			or planet.roads.has_facility(node) \
+			or planet.planet_data.get_node(node).is_occupied:
 		return [false, "Land is already occupied"]
 	
 	if not node in city.cells:
@@ -137,14 +150,14 @@ func can_set_land_use(city: City, node: int, land_use: int, consider_veg_res: bo
 	if not consider_veg_res:
 		return [true, null]
 	
-	var veg = planet_data.get_node(node).vegetation_type
+	var veg = planet.planet_data.get_node(node).vegetation_type
 	var lu: Dictionary = constants.LU_MAPPING[land_use]
 	if not veg in lu:
 		return [false, "Land use %s not possible on %s" % [LandUse.LU_NAMES[land_use], LandUse.VEG_NAMES[veg]]]
 	
 	var extract_resource = LandUse.LU_RESOURCE[land_use]
 	if extract_resource != null:
-		var res_here = resources.resources.get(node, null)
+		var res_here = planet.resources.resources.get(node, null)
 		if res_here == null or res_here[0] != extract_resource:
 			return [false, "Land use %s not possible without resource %s" % [LandUse.LU_NAMES[land_use], Resources.RES_NAMES[extract_resource]]]
 	
@@ -155,8 +168,8 @@ func set_land_use(city: City, node: int, land_use: int):
 	if land_use == LandUse.LU_NONE:
 		if node in city.land_use:
 			city.clear_land_use(node)
-			planet_data.set_occupied(node, false)
-			city.update_visuals(planet_data)
+			planet.planet_data.set_occupied(node, false)
+			city.update_visuals(planet.planet_data)
 			return null
 		else:
 			return "Can't clear, land is not is use"
@@ -168,8 +181,8 @@ func set_land_use(city: City, node: int, land_use: int):
 	if city.workers() < LandUse.LU_WORKERS[land_use]:
 		return "Not enough workers (requires %d)" % LandUse.LU_WORKERS[land_use]
 	
-	city.set_land_use(planet_data, node, land_use)
-	planet_data.set_occupied(node, true)
-	city.update_visuals(planet_data)
+	city.set_land_use(planet.planet_data, node, land_use)
+	planet.planet_data.set_occupied(node, true)
+	city.update_visuals(planet.planet_data)
 	
 	return null