+++ scripts/managers/road_network.gd
@@ -113,13 +113,12 @@ func disconnect_points(v1: int, v2: int):
 	_disconnect(v2, v1)
 
 
-func add_facility(v: int, facility: Facility):
-	assert(not facilities.has(v), "There is already a facility at node %s" % v)
-	facilities[v] = facility
+func add_facility(node: int, facility: Facility):
+	assert(not facilities.has(node), "There is already a facility at node %s" % node)
+	facilities[node] = facility
 
-
-func remove_facility(v: int):
-	assert(facilities.erase(v), "There is no a facility at node %s to remove" % v)
+func remove_facility(node: int):
+	assert(facilities.erase(node), "There is no a facility at node %s to remove" % node)
 
 
 func has_facility(v: int) -> bool: